Event-Handler Code Fixed
Fixed event-handler Javascript code. The extension would automatically fire off the infinite loop immediately when the user navigated to a Power BI Report's URL, even before the extension was initialized by the user. Now, the extension will not fire off the infinite loop until the extension is actually initialized by the user.

This was a great learning experience for me, as the problem was in understanding asynchronous function calling and callback functions. I am used to procedural programming, but now I have experience and knowledge in how to "create" a procedural outline for asynchronous functions via callback functions! I'm excited to see what else I can accomplish with utilizing these Javascript libraries.

I plan to return to this project again soon and allow users the ability to stop the extension after it has been initialized, as that feature has not yet been added.