End of Day
I successfully incorporated the accumulation variable concept to only make the extension refresh once a certain threshold of slide times have been reached. This will moderate performance and load on the data sources by limiting the number of times the refresh button is clicked by the extension.

This is the best option for the moment, since vanilla Javascript is a singe-threaded language. To get the TRULY functional version of this concept that I originally planned, I would need to incorporate additional Javascript libraries for multi-threaded processing. This will suffice for the moment.